+++ docs/CHANGELOG.md
@@ -1,7 +1,7 @@
 * 2026/08/03
   - Updated error message expectation to updates in PJRT.
   - Bumped XLA PJRT version to v0.114.0
-  - Exposed `pjrt.IsCudaName`
+  - Exposed `pjrt.IsCUDAName`
   - Set default "preallocate" option only for CUDA backends.
 
 # v0.3.0: API changes for GoMLX v0.28.0 and gomlx/compute v0.1.0; Added flash-attention for CUDA.